pyproject.toml
[project]
name = "vandelay"
version = "0.1.0"
description = "Vandelay — the employee who doesn't exist. An always-on AI agent platform powered by Agno."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Shaun", email = "" }]
keywords = ["ai", "agent", "agno", "vandelay", "assistant", "automation", "always-on"]

dependencies = [
    "agno>=1.5",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "typer>=0.15",
    "questionary>=2.1",
    "rich>=13.9",
    "pydantic-settings>=2.7",
    "httpx>=0.28",
    "apscheduler>=3.11",
    "playwright>=1.49",
    "lancedb>=0.17",
    "croniter>=6.0",
    "python-telegram-bot>=21",
    "sqlalchemy>=2.0.46",
    "anthropic>=0.79.0",
    "pyjwt>=2.11.0",
    "nodeenv>=1.9",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.60"]
google = ["google-genai>=1.0"]
ollama = ["ollama>=0.4"]
postgres = ["psycopg[binary]>=3.2"]
all = [
    "anthropic>=0.40",
    "openai>=1.60",
    "google-genai>=1.0",
    "ollama>=0.4",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "ruff>=0.9",
]

[project.scripts]
vandelay = "vandelay.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vandelay"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
